added proxy option
`pre-build.js` tries to download the required files from the server but if
you are behind a corporate proxy server it will fail. By using the
environment variables to check for a proxy setting this could be avoided
and all those sad little men behind a proxy will be happy.